# tests: Add benchmark compare script and include in CI