**What this PR does / why we need it**:

Add a script that runs all benchmark tests on the current git ref and
the last release and add it to the Makefile and .travis.yml file.

This does not enforce certain benchmark results, but at least gives people the chance to easily compare their changes performance-wise.

Idea came up here: https://github.com/kubernetes/kube-state-metrics/pull/633#issuecomment-458084905