@brancz 2019-01-28
The CI job currently reports:

```
$ make test-benchmark-compare
make: Nothing to be done for `test-benchmark-compare'.
```