@mxinden 2019-01-28
This looks better. Sorry about that @brancz. Had some difficulties making remote branches available in the build (Thanks @metalmatze).

```
./tests/compare_benchmarks.sh release-1.5
### Testing HEAD
goos: linux
goarch: amd64
pkg: k8s.io/kube-state-metrics
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1000873656 ns/op
BenchmarkKubeStateMetrics/MakeRequests-2            	1000000000	         0.09 ns/op	357547622462.10 MB/s
PASS
ok  	k8s.io/kube-state-metrics	3.035s
PASS
ok  	k8s.io/kube-state-metrics/internal/collector	0.012s
?   	k8s.io/kube-state-metrics/pkg/collector	[no test files]
?   	k8s.io/kube-state-metrics/pkg/constant	[no test files]
goos: linux
goarch: amd64
pkg: k8s.io/kube-state-metrics/pkg/metric
BenchmarkMetricWrite/value-1-2         	 2000000	       804 ns/op
BenchmarkMetricWrite/value-35.7-2      	 2000000	       975 ns/op
PASS
ok  	k8s.io/kube-state-metrics/pkg/metric	5.368s
PASS
ok  	k8s.io/kube-state-metrics/pkg/metrics_store	0.007s
PASS
ok  	k8s.io/kube-state-metrics/pkg/options	0.007s
?   	k8s.io/kube-state-metrics/pkg/version	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/pkg/whiteblacklist	0.003s
PASS
ok  	k8s.io/kube-state-metrics/tests/lib	0.013s
### Done testing HEAD
### Testing release-1.5
Warning: you are leaving 2 commits behind, not connected to
any of your branches:
  f29e681 Merge f4e79baa24d332d6533949158d5ff8ddf4b7bb71 into 5b394f851d1d4d456b4ee8ffe0c86dc0a752e4b6
  f4e79ba tests: Add benchmark compare script and include in CI
If you want to keep them by creating a new branch, this may be a good time
to do so with:
 git branch <new-branch-name> f29e681
Branch 'release-1.5' set up to track remote branch 'release-1.5' from 'origin'.
Switched to a new branch 'release-1.5'
goos: linux
goarch: amd64
pkg: k8s.io/kube-state-metrics
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1000872211 ns/op
BenchmarkKubeStateMetrics/MakeRequests-2            	1000000000	         0.09 ns/op	354413869848.21 MB/s
PASS
ok  	k8s.io/kube-state-metrics	3.019s
PASS
ok  	k8s.io/kube-state-metrics/pkg/collectors	0.013s
?   	k8s.io/kube-state-metrics/pkg/constant	[no test files]
goos: linux
goarch: amd64
pkg: k8s.io/kube-state-metrics/pkg/metrics
BenchmarkMetricWrite/value-1-2         	 2000000	       666 ns/op
BenchmarkMetricWrite/value-35.7-2      	 2000000	       854 ns/op
PASS
ok  	k8s.io/kube-state-metrics/pkg/metrics	4.582s
PASS
ok  	k8s.io/kube-state-metrics/pkg/metrics_store	0.008s
PASS
ok  	k8s.io/kube-state-metrics/pkg/options	0.006s
?   	k8s.io/kube-state-metrics/pkg/version	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/pkg/whiteblacklist	0.003s
PASS
ok  	k8s.io/kube-state-metrics/tests/lib	0.014s
### Done testing release-1.5
Note: checking out 'f29e68138530ee640ff33b49410f8e57f5128961'.
You are in 'detached HEAD' state. You can look around, make experimental
changes and commit them, and you can discard any commits you make in this
state without impacting any branches by performing another checkout.
If you want to create a new branch to retain commits you create, you may
do so (now or later) by using -b with the checkout command again. Example:
  git checkout -b <new-branch-name>
HEAD is now at f29e681... Merge f4e79baa24d332d6533949158d5ff8ddf4b7bb71 into 5b394f851d1d4d456b4ee8ffe0c86dc0a752e4b6
### Result
old=release-1.5 new=HEAD
benchmark                                       old ns/op      new ns/op      delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     1000872211     1000873656     +0.00%
BenchmarkKubeStateMetrics/MakeRequests-2        0.09           0.09           +0.00%
BenchmarkMetricWrite/value-1-2                  666            804            +20.72%
BenchmarkMetricWrite/value-35.7-2               854            975            +14.17%
benchmark                                    old MB/s            new MB/s            speedup
BenchmarkKubeStateMetrics/MakeRequests-2     354413869848.21     357547622462.10     1.01x
```