@mxinden 2019-01-28
Also added memory benchmarks:

```
### Result
old=release-1.5 new=HEAD
benchmark                                       old ns/op      new ns/op      delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     1008699032     1000892821     -0.77%
BenchmarkKubeStateMetrics/MakeRequests-2        0.09           0.09           +0.00%
BenchmarkMetricWrite/value-1-2                  674            821            +21.81%
BenchmarkMetricWrite/value-35.7-2               855            1006           +17.66%
benchmark                                    old MB/s            new MB/s            speedup
BenchmarkKubeStateMetrics/MakeRequests-2     356252143551.77     351465698658.81     0.99x
benchmark                                       old allocs     new allocs     delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     730825         731751         +0.13%
BenchmarkKubeStateMetrics/MakeRequests-2        0              0              +0.00%
BenchmarkMetricWrite/value-1-2                  5              7              +40.00%
BenchmarkMetricWrite/value-35.7-2               5              7              +40.00%
benchmark                                       old bytes     new bytes     delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     66243336      65665408      -0.87%
BenchmarkKubeStateMetrics/MakeRequests-2        0             0             +0.00%
BenchmarkMetricWrite/value-1-2                  512           536           +4.69%
BenchmarkMetricWrite/value-35.7-2               512           536           +4.69%
```

I will need to look into the new allocations before the next release. But I think that should happen in a follow up PR.