@brancz 2019-01-29
nice, yes this is exactly what I was thinking :+1: 

/lgtm
/approve